// spring-ai-web/src/app/@core/api/generated/models/chatbot-resource.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ChatbotStatusEnum } from '../models/chatbot-status-enum';
import { OrganizationResource } from '../models/organization-resource';
export interface ChatbotResource {
  confidenceThreshold?: number;
  customGuardrails?: string;
  description?: string;
  id?: string;
  metadataFilters?: {
[key: string]: string;
};
  name?: string;
  organization?: OrganizationResource;
  responseTemplate?: string;
  status?: ChatbotStatusEnum;
  strictGuardrails?: boolean;
  systemPromptTemplate?: string;
  temperature?: number;
}